refactor: 파일 구조 정리
- app.controller/service 삭제 (NestJS 기본 생성 껍데기)
- notifications.repository → repositories/ 서브폴더로 이동
- auth/types/, notifications/types/ 폴더 제거 후 상위로 이동
- GroupType type alias → enum으로 변환 및 games/enums/로 이동
- ParseEnumPipe(['college', 'department']) → ParseEnumPipe(GroupType)
