+++ src/games/enums/group-type.enum.ts
@@ -1,0 +1,4 @@
+export enum GroupType {
+  COLLEGE = 'college',
+  DEPARTMENT = 'department',
+}
