--- src/games/types/group-type.type.ts
@@ -1,1 +1,0 @@
-export type GroupType = 'college' | 'department';
